New houserules, Scala 2.11 & scalac opts.

+++ io/src/main/scala/sbt/internal/io/SourceModificationWatch.scala
@@ -30,7 +30,7 @@ private[sbt] object SourceModificationWatch {
       if (triggered)
         (true, newState)
       else {
-        Thread.sleep(pollDelayMillis)
+        Thread.sleep(pollDelayMillis.toLong)
         if (terminationCondition)
           (false, newState)
         else
